discovery: refactor gh provider iterator
